requirements.txt
Django==1.8.2
psycopg2==2.6
djangorestframework==3.3.1
markdown==2.6.4
django-filter==0.11.0
django-rest-framework-social-oauth2==1.0.2